% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_rpath.R
\name{fn_rpath}
\alias{fn_rpath}
\title{Computes long-term equilibrium yields for \code{Rpath} models}
\usage{
fn_rpath(
  par,
  simul.years = 100,
  aged.str,
  data.years,
  IDnames,
  rsim.mod,
  rpath.params,
  avg.window = 10,
  integration.method = "RK4"
)
}
\arguments{
\item{integration.method}{}
}
\value{

}
\description{
Description
}
