% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nash.R
\name{nash}
\alias{nash}
\title{Computes Nash Equilibrium harvesting rates}
\usage{
nash(
  par,
  fn,
  ...,
  method = "LV",
  F.increase = 0.1,
  yield.curves = TRUE,
  conv.criterion = 0.001
)
}
\arguments{
\item{conv.criterion}{}
}
\value{

}
\description{
Function that (i) takes harvesting values; (ii) reparameterises
the model; (iii) simulates forward in time to equilibrium returning
yields for commercially-important species averaged over a certain
time window.
}
