% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nash.R
\name{nash}
\alias{nash}
\title{Computes harvesting rates at the Nash Equilibrium}
\usage{
nash(
  par,
  fn,
  ...,
  method = "LV",
  yield.curves = FALSE,
  conv.criterion = 0.001,
  Bcons = 0,
  F.increase = 0.1,
  progress = TRUE
)
}
\arguments{
\item{par}{Numeric vector of harvesting rates of length equal to the number of
harvested species.}

\item{fn}{Function that runs the ODE model with \code{par} as input and
returns simulated yields at equilibrium.}

\item{...}{Further arguments to be passed to \code{fn}.}

\item{method}{Method utilised to compute the Nash Equilibrium:
(i) `\code{LV}` or (ii) `\code{round-robin}` method.}

\item{conv.criterion}{Absolute convergence tolerance set by default to
\eqn{< 0.001}.}

\item{Bcons}{Biomass conservation threshold. Default set to \eqn{0}.}

\item{F.increase}{Double type numeric vector indicating the step size used
to compute the interaction matrix.}

\item{yield.cruves}{Logical TRUE/FALSE if equilibrium yield curves for each
species are desired.}
}
\value{
The function \code{nash} returns a list with the following components:
\item{par}{Harvesting rates at the Nash Equilibrium.}
\item{value}{Value of \code{fn} corresponding to the optimised \code{par}.}
\item{counts}{Number of calls to \code{fn}.}
\item{convergence}{Statement indicating the number of iterations for
\code{conv.criterion} to be reached.}
}
\description{
Function that takes a numeric vector `\code{par}` of initial harvesting rates
to be optimised by evaluating the objective function `\code{fn}` that is to
be maximised. Its structure in terms of arguments and outputs is analogous
to that of \code{\link{optim}}.
}
\details{

}
\references{
\insertRef{Matsuda2006}{nash}

\insertRef{Worm2009}{nash}

\insertRef{Thorpe2017}{nash}

\insertRef{Pope2019}{nash}

\insertRef{Farcas2016}{nash}

\insertRef{Lucey2020}{nash}
}
