% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nash.R
\name{nash}
\alias{nash}
\title{Computes harvesting rates at the Nash Equilibrium (NE)}
\usage{
nash(
  par,
  fn,
  ...,
  method = "LV",
  yield.curves = FALSE,
  conv.criterion = 0.001,
  Bcons = 0,
  F.increase = 0.1,
  progress = TRUE,
  track = FALSE
)
}
\arguments{
\item{par}{Double type numeric vector of harvesting rates of length equal to
the number of harvested species for which the NE is desired.}

\item{fn}{Function that runs the multispecies/ecosystem model with \code{par}
as input and returns simulated yields at equilibrium.}

\item{...}{Further arguments to be passed to \code{fn}.}

\item{method}{Method utilised to compute the Nash Equilibrium:
(i) `\code{LV}` or (ii) `\code{round-robin}` method (see `Details` section
for specifics).}

\item{yield.curves}{Logical TRUE/FALSE if equilibrium yield curves for each of
the optimised species are to be computed.}

\item{conv.criterion}{Absolute convergence tolerance set by default to
\eqn{< 0.001}.}

\item{Bcons}{Constraints for biodiversity conservation. Double type numeric
vector set to \eqn{0} by default.}

\item{F.increase}{Double type numeric vector indicating the step size used
to compute the effective interaction matrix \eqn{M}.}

\item{progress}{Logical that if TRUE information on the progress of the
optimisation is produced.}

\item{track}{Logical that if TRUE will return all fishing mortality rates that
where iteratively computed during the search.

\loadmathjax}
}
\value{
The function \code{nash} returns a list with the following components:
\item{par}{Harvesting rates at the NE.}
\item{value}{Yield values of \code{fn} corresponding to the optimised
\code{par}.}
\item{counts}{Number of \code{fn} evaluations until NE.}
\item{convergence}{Statement indicating the number of iterations for
\code{conv.criterion} to be reached.}
}
\description{
Takes a numeric vector of initial harvesting rates to be optimised
(`\code{par}`) by evaluating the objective function that is to be maximised
(`\code{fn}`). Its structure in terms of arguments and outputs is similar
to that of \code{\link{optim}}.
}
\details{
For ecosystem models where there is some interest in keeping some or
all harvested species above a certain biomass state limit, \code{Bcons}
should be populated with non-zero biomass values. The length of this vector
must be the same as \code{par} and set to non-zero where relevant. In the
literature, is common practice to fixed such \emph{constraints of
biodiversity conservation} \insertCite{Matsuda2006}{nash} as a fraction
(\emph{e.g.} \mjeqn{0.1-0.25}{ascii}) of the unfished biomass
\mjeqn{B_0}{ascii}; biomass threshold at which a stock is considered
collapsed \insertCite{Worm2009}{nash}.

Equilibrium yield curves are obtained for each \eqn{i} species by applying
different harvesting values to \eqn{i} whilst keeping the other species
\eqn{j} at the optimised \code{par} levels (\emph{i.e.} at
\mjeqn{\mathbf{F_{Nash}}}{ascii}).
The harvesting values applied to \eqn{i} run from \eqn{0}
to \mjeqn{F_{Nash,i}\times 2}{ascii} with a desired sequence length
of \code{length.out}\eqn{=30} (see \code{\link{seq}} for details).
As raised by \insertCite{Thorpe2017;textual}{nash}, this is one of the
advantages of using the NE as the multispecies extension of the
\emph{Maximum Sustainable Yield} concept.

To compute the interaction matrix a second order central difference quotient
is used to approximate derivatives. \code{F.increase} is employed during
this calculation as a step-size set by default to \eqn{0.1} to avoid
truncation and/or rounding errors \insertCite{Pope2019}{nash}.

The `\code{LV}` method is set by default given its performance advantage
\insertCite{Oneill2023}{nash}
over the `\code{round-robin}` method and is based on the protocol devised by
\insertCite{Farcas2016}{nash}. For each species \eqn{i} in turn,
\code{round-robin} iteratively maximises the yield by adjusting the harvesting
rates whereas \code{LV} does the same simultaneously for all species per
iteration.
}
\references{
\insertRef{Farcas2016}{nash}

\insertRef{Lucey2020}{nash}

\insertRef{Matsuda2006}{nash}

\insertRef{Oneill2023}{nash}

\insertRef{Pope2019}{nash}

\insertRef{Thorpe2017}{nash}
}
